creating new hook for autoprune requests and adding check that user cant submit 0 for value
