+++ web/src/hooks/UseOrganization.ts
@@ -1,12 +1,5 @@
-import {
-  NamespaceAutoPrunePolicy,
-  createNamespaceAutoPrunePolicy,
-  deleteNamespaceAutoPrunePolicy,
-  fetchNamespaceAutoPrunePolicies,
-  fetchOrg,
-  updateNamespaceAutoPrunePolicy,
-} from 'src/resources/OrganizationResource';
-import {useMutation, useQuery, useQueryClient} from '@tanstack/react-query';
+import {fetchOrg} from 'src/resources/OrganizationResource';
+import {useQuery} from '@tanstack/react-query';
 import {useOrganizations} from './UseOrganizations';
 
 export function useOrganization(name: string) {
@@ -28,124 +21,5 @@ export function useOrganization(name: string) {
     error,
     loading: isLoading,
     organization,
-  };
-}
-
-export function useNamespaceAutoPrunePolicies(
-  namespace: string,
-  isUser: boolean,
-) {
-  const {
-    data: policies,
-    isLoading,
-    error,
-    isSuccess,
-    dataUpdatedAt,
-  } = useQuery(['namespace', 'autoprunepolicies', namespace], ({signal}) =>
-    fetchNamespaceAutoPrunePolicies(namespace, isUser, signal),
-  );
-
-  return {
-    error,
-    isSuccess,
-    isLoading,
-    dataUpdatedAt,
-    policies,
-  };
-}
-
-export function useCreateNamespaceAutoPrunePolicy(
-  namespace: string,
-  isUser: boolean,
-) {
-  const queryClient = useQueryClient();
-  const {
-    mutate: createPolicy,
-    isSuccess: successCreatePolicy,
-    isError: errorCreatePolicy,
-    error: errorCreatePolicyDetails,
-  } = useMutation(
-    async (policy: NamespaceAutoPrunePolicy) =>
-      createNamespaceAutoPrunePolicy(namespace, policy, isUser),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries([
-          'namespace',
-          'autoprunepolicies',
-          namespace,
-        ]);
-      },
-    },
-  );
-
-  return {
-    createPolicy: createPolicy,
-    successCreatePolicy: successCreatePolicy,
-    errorCreatePolicy: errorCreatePolicy,
-    errorCreatePolicyDetails: errorCreatePolicyDetails,
-  };
-}
-
-export function useUpdateNamespaceAutoPrunePolicy(
-  namespace: string,
-  isUser: boolean,
-) {
-  const queryClient = useQueryClient();
-  const {
-    mutate: updatePolicy,
-    isSuccess: successUpdatePolicy,
-    isError: errorUpdatePolicy,
-    error: errorUpdatePolicyDetails,
-  } = useMutation(
-    async (policy: NamespaceAutoPrunePolicy) =>
-      updateNamespaceAutoPrunePolicy(namespace, policy, isUser),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries([
-          'namespace',
-          'autoprunepolicies',
-          namespace,
-        ]);
-      },
-    },
-  );
-
-  return {
-    updatePolicy: updatePolicy,
-    successUpdatePolicy: successUpdatePolicy,
-    errorUpdatePolicy: errorUpdatePolicy,
-    errorUpdatePolicyDetails: errorUpdatePolicyDetails,
-  };
-}
-
-export function useDeleteNamespaceAutoPrunePolicy(
-  namespace: string,
-  isUser: boolean,
-) {
-  const queryClient = useQueryClient();
-  const {
-    mutate: deletePolicy,
-    isSuccess: successDeletePolicy,
-    isError: errorDeletePolicy,
-    error: errorDeletePolicyDetails,
-  } = useMutation(
-    async (uuid: string) =>
-      deleteNamespaceAutoPrunePolicy(namespace, uuid, isUser),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries([
-          'namespace',
-          'autoprunepolicies',
-          namespace,
-        ]);
-      },
-    },
-  );
-
-  return {
-    deletePolicy: deletePolicy,
-    successDeletePolicy: successDeletePolicy,
-    errorDeletePolicy: errorDeletePolicy,
-    errorDeletePolicyDetails: errorDeletePolicyDetails,
   };
 }
